feat: 全量修复三端对齐问题 F1-F15
Database:
- F1: PG/MySQL init 脚本合并 RuoYi sys_* 表 (11 张)
- F2: 业务 12 张表补齐 create_by/update_by 列

Backend:
- F15: PermissionScope 响应拆分 modulesText -> modules 数组

Frontend:
- F3: 新增通知管理模块 (types/api/view/route/menu)
- F5: 统一 API 返回类型 PageResponse<T>/ApiResponse<T>
- F9: 删除 api/user.ts 死代码
- F11: 移除测试模块空壳

Desktop:
- F6: ToolProject.RunMode 统一为 string (ToolRunModes 静态类)
- F7: User.Id/Game.Id int -> long
- F12: 引入 appsettings.json 配置化 BaseURL

Docs:
- F4: 重写 docs/database-design.md, 30 张表完整对齐

+++ desktop/ZlinksPackageSystem.Desktop/Models/Game.cs
@@ -4,7 +4,10 @@ namespace ZlinksPackageSystem.Desktop.Models
 {
     public class Game
     {
-        public int Id { get; set; }
+        /// <summary>
+        /// 游戏 ID(与后端 Java Long 对齐,避免 > 2^31 时的精度丢失)
+        /// </summary>
+        public long Id { get; set; }
         public string GameName { get; set; } = string.Empty;
         public string GameDirection { get; set; } = string.Empty;
         public string Source { get; set; } = string.Empty;
